requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Acquisition
alpha-vantage==2.3.1
tiingo==0.14.0
yfinance==0.2.35
pandas-datareader==0.10.0
fredapi==0.5.1
finnhub-python==2.4.19
praw==7.7.1  # Reddit API
polygon-api-client==1.12.5
openbb==3.2.4

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
timescaledb==0.0.4
alembic==1.13.1

# Caching & Messaging
redis==5.0.1
redis-py-cluster==2.1.3
kafka-python==2.0.2
confluent-kafka==2.3.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
numba==0.58.1
polars==0.20.3  # High-performance alternative to pandas

# Machine Learning - Core
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2
optuna==3.5.0  # Hyperparameter optimization

# Deep Learning
torch==2.1.2
tensorflow==2.15.0
keras==2.15.0
transformers==4.36.2  # For FinBERT
huggingface-hub==0.20.2

# ML Optimization
onnx==1.15.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
intel-extension-for-sklearn==2023.2.1  # daal4py
openvino==2023.3.0

# NLP
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml==4.9.3

# Financial Libraries
quantlib==1.32
ffn==0.3.7
empyrical==0.5.5
pyfolio==0.9.2
bt==0.2.9  # Backtesting framework
vectorbt==0.26.2
zipline-reloaded==2.4.2

# Portfolio Optimization
riskfolio-lib==5.1.0
PyPortfolioOpt==1.5.5
cvxpy==1.4.1
cvxopt==1.3.2

# Technical Analysis
ta==0.11.0
ta-lib==0.4.28
pandas-ta==0.3.14b0
finta==1.3

# Statistics & Econometrics
statsmodels==0.14.1
arch==6.2.0  # GARCH models
hmmlearn==0.3.0  # Hidden Markov Models
sklearn-genetic==0.5.1

# Interactive Brokers
ib-insync==0.9.86
ibapi==9.81.1

# FIX Protocol
quickfix==1.15.1
simplefix==1.0.16

# Risk Management
QuantLib-Python==1.32
pykalman==0.9.5

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-exporter-prometheus==0.43b0
opentelemetry-exporter-jaeger==1.22.0
prometheus-client==0.19.0
grafana-api==1.0.3

# Utilities
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
tenacity==8.2.3  # Retry logic
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
joblib==1.3.2
dask==2024.1.0
ray==2.9.0  # Distributed computing

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.39.2
colorlog==6.8.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
bokeh==3.3.2
dash==2.14.2

# Testing & Quality
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
pytest-mock==3.12.0
pytest-benchmark==4.0.0
hypothesis==6.92.1  # Property-based testing
faker==21.0.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
bandit==1.7.5  # Security linting

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3
mkdocs-material==9.5.3

# Development Tools
ipython==8.19.0
jupyter==1.0.0
jupyterlab==4.0.10
notebook==7.0.6
ipywidgets==8.1.1

# Deployment
gunicorn==21.2.0
uvicorn==0.25.0
fastapi==0.108.0
celery==5.3.4
flower==2.0.1  # Celery monitoring

# Security
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# Feature Store
feast==0.35.0

# ML Experiment Tracking
mlflow==2.9.2
wandb==0.16.2
neptune-client==1.9.1

# Data Validation
great-expectations==0.18.8
pandera==0.17.2
pydantic-extra-types==2.3.0

# Alternative Data
gdelt==0.1.11  # GDELT news database
tweepy==4.14.0  # Twitter API (legacy support)

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0
line-profiler==4.1.1

# Miscellaneous
click==8.1.7  # CLI framework
rich==13.7.0  # Terminal formatting
tabulate==0.9.0
